Find the most amount of calories an elf is carrying
Approach:
Since we just care about the max amount, we iterativly
go through the input line by line. We keep a running count
of the current elf's calories by adding each food item's
calorie amount until we see a line break. At the line break
we compare the running count to the max, if it's greater than
we update the max with the running count, if not then we
continue onto the next elf starting from a running count of 0.